+++ generate/src/MonologueComposition.tsx
@@ -1,4 +1,3 @@
-import { useAudioData, visualizeAudio } from '@remotion/media-utils';
 import React, { useEffect, useRef, useState } from 'react';
 import {
 	AbsoluteFill,
@@ -13,237 +12,12 @@ import {
 } from 'remotion';
 import { fps, music } from './tmp/context';
 import { PaginatedSubtitles } from './Subtitles';
-import { z } from 'zod';
-import { zColor } from '@remotion/zod-types';
-
-export type WordTiming = {
-	word: string;
-	start: number;
-	end: number;
-};
-
-export type SubtitleEntry = {
-	index: string;
-	startTime: number;
-	endTime: number;
-	text: string;
-	srt: string;
-	srtFileIndex: number;
-	wordTimings: WordTiming[];
-};
-
-const AgentDetailsSchema = z.record(
-	z.object({
-		color: zColor(),
-		image: z.string().refine((s) => s.endsWith('.png'), {
-			message: 'Agent image must be a .png file',
-		}),
-	})
-);
-
-const srtTimeToSeconds = (srtTime: string) => {
-	const [hours, minutes, secondsAndMillis] = srtTime.split(':');
-	const [seconds, milliseconds] = secondsAndMillis.split(',');
-	return (
-		Number(hours) * 3600 +
-		Number(minutes) * 60 +
-		Number(seconds) +
-		Number(milliseconds) / 1000
-	);
-};
-
-const parseSRT = (
-	srtContent: string,
-	srtFileIndex: number
-): SubtitleEntry[] => {
-	const blocks = srtContent.split('\n\n');
-	const MIN_DURATION = 0.5;
-
-	const preliminaryEntries = blocks
-		.map((block) => {
-			const lines = block.split('\n');
-			const indexLine = lines[0];
-			const timeLine = lines[1];
-
-			if (!indexLine || !timeLine || lines.length < 3) {
-				return null;
-			}
-
-			const [startTime, endTime] = timeLine
-				.split(' --> ')
-				.map(srtTimeToSeconds);
-
-			const textLines = lines.slice(2).join(' ');
-
-			// Calculate word timings
-			const words = textLines.split(' ');
-			const timePerWord = (endTime - startTime) / words.length;
-			const wordTimings = words.map((word, idx) => ({
-				word,
-				start: startTime + idx * timePerWord,
-				end: startTime + (idx + 1) * timePerWord,
-			}));
-
-			return {
-				index: indexLine,
-				startTime,
-				endTime,
-				text: textLines,
-				srt: block,
-				srtFileIndex,
-				wordTimings,
-			};
-		})
-		.filter((entry): entry is SubtitleEntry => entry !== null);
-
-	const combinedEntries: SubtitleEntry[] = [];
-	let currentEntry: SubtitleEntry | null = null;
-	let accumulatedText: string[] = [];
-	let accumulatedWordTimings: WordTiming[] = [];
-
-	for (const entry of preliminaryEntries) {
-		if (!currentEntry) {
-			currentEntry = entry;
-			accumulatedText = [entry.text];
-			accumulatedWordTimings = [...entry.wordTimings];
-			continue;
-		}
-
-		const currentDuration = currentEntry.endTime - currentEntry.startTime;
-
-		if (currentDuration < MIN_DURATION) {
-			accumulatedText.push(entry.text);
-			accumulatedWordTimings.push(...entry.wordTimings);
-			currentEntry = {
-				...currentEntry,
-				endTime: entry.endTime,
-				text: accumulatedText.join(' '),
-				wordTimings: accumulatedWordTimings,
-				srt: `${currentEntry.index}\n${secondsToSrtTime(
-					currentEntry.startTime
-				)} --> ${secondsToSrtTime(entry.endTime)}\n${accumulatedText.join(
-					' '
-				)}`,
-			};
-		} else {
-			combinedEntries.push(currentEntry);
-			currentEntry = entry;
-			accumulatedText = [entry.text];
-			accumulatedWordTimings = [...entry.wordTimings];
-		}
-	}
-
-	if (currentEntry) {
-		combinedEntries.push(currentEntry);
-	}
-
-	return combinedEntries;
-};
-
-const secondsToSrtTime = (seconds: number): string => {
-	const hours = Math.floor(seconds / 3600);
-	const minutes = Math.floor((seconds % 3600) / 60);
-	const secs = Math.floor(seconds % 60);
-	const millis = Math.floor((seconds % 1) * 1000);
-
-	return `${String(hours).padStart(2, '0')}:${String(minutes).padStart(
-		2,
-		'0'
-	)}:${String(secs).padStart(2, '0')},${String(millis).padStart(3, '0')}`;
-};
-
-const SubtitleFileSchema = z.object({
-	name: z.string(),
-	file: z.string().refine((s) => s.endsWith('.srt'), {
-		message: 'Subtitle file must be a .srt file',
-	}),
-	asset: z.string(),
-});
-
-export const AudioGramSchema = z.object({
-	initialAgentName: z.string(),
-	agentDetails: AgentDetailsSchema,
-	videoFileName: z.string().optional(),
-	useBackground: z.boolean(),
-	durationInSeconds: z.number().positive(),
-	audioOffsetInSeconds: z.number().min(0),
-	subtitlesFileName: z.array(SubtitleFileSchema),
-	audioFileName: z.string().refine((s) => s.endsWith('.mp3'), {
-		message: 'Audio file must be a .mp3 file',
-	}),
-	titleText: z.string(),
-	titleColor: zColor(),
-	subtitlesTextColor: zColor(),
-	subtitlesLinePerPage: z.number().int().min(0),
-	subtitlesLineHeight: z.number().int().min(0),
-	subtitlesZoomMeasurerSize: z.number().int().min(0),
-	mirrorWave: z.boolean(),
-	waveLinesToDisplay: z.number().int().min(0),
-	waveFreqRangeStartIndex: z.number().int().min(0),
-	waveNumberOfSamples: z.enum(['32', '64', '128', '256', '512']),
-});
-
-type AudiogramCompositionSchemaType = z.infer<typeof AudioGramSchema>;
-
-const AudioViz: React.FC<{
-	numberOfSamples: number;
-	freqRangeStartIndex: number;
-	waveColor: string;
-	waveLinesToDisplay: number;
-	mirrorWave: boolean;
-	audioSrc: string;
-}> = ({
-	numberOfSamples,
-	waveColor,
-	freqRangeStartIndex,
-	waveLinesToDisplay,
-	mirrorWave,
-	audioSrc,
-}) => {
-	const frame = useCurrentFrame();
-
-	const audioData = useAudioData(audioSrc);
-
-	if (!audioData) {
-		return null;
-	}
-
-	const frequencyData = visualizeAudio({
-		fps,
-		frame,
-		audioData,
-		numberOfSamples,
-	});
-
-	const frequencyDataSubset = frequencyData.slice(
-		freqRangeStartIndex,
-		freqRangeStartIndex +
-			(mirrorWave ? Math.round(waveLinesToDisplay / 2) : waveLinesToDisplay)
-	);
-
-	const frequencesToDisplay = mirrorWave
-		? [...frequencyDataSubset.slice(1).reverse(), ...frequencyDataSubset]
-		: frequencyDataSubset;
-
-	return (
-		<div className="transition-all audio-viz z-30">
-			{frequencesToDisplay.map((v, i) => {
-				return (
-					<div
-						key={i}
-						className={`z-30 bar `}
-						style={{
-							backgroundColor: waveColor,
-							minWidth: '1px',
-							opacity: 0.5,
-							height: `${500 * Math.sqrt(v)}%`,
-						}}
-					/>
-				);
-			})}
-		</div>
-	);
-};
+import {
+	AudioViz,
+	SubtitleEntry,
+	AudiogramCompositionSchemaType,
+	parseSRT,
+} from './composition_helpers';
 
 export const MonologueComposition: React.FC<AudiogramCompositionSchemaType> = ({
 	subtitlesFileName,
